+++ lib/main.dart
@@ -1,26 +1,13 @@
-import 'package:flutter/material.dart';
-//paquete basico para proyectos Flutter
+import 'package:flutter/material.dart'; //paquete basico para proyectos Flutter
+
+//Nosotros escribimos las importaciones nuestas al final
+//import 'src/app.dart'; Forma 1 importación
+import 'package:contador/src/app.dart'; //Forma 2 importación
+
 void main() { //Función principal de la aplicación
 
- runApp(new MyApp()); //metodo de punto de entrada de Aplicación
+ runApp( MyApp()); //metodo de punto de entrada de Aplicación
 
 
 }
 
-
-
-class MyApp extends StatelessWidget {
-
-  @override
-  build( context ) { //context contiene información del arbol de widgets, conoce relación padre-hijo
-    //Doc info: https://api.flutter.dev/flutter/material/MaterialApp-class.html
-    return MaterialApp(
-      home: Center(
-        child: Text('Hola Mundo'),
-        )      
-    ); 
-
-  }
-
-
-}
